[refactor] 테스트 케이스 코드 통합
